+++ TinkoffChat/1. PresentationLayer/Modules/Extensions/UIImage.swift
