+++ TinkoffChat/1. PresentationLayer/Modules/General/ChannelListViewController/Model/ChannelList+UITableViewDelegate.swift
