+++ TinkoffChat/1. PresentationLayer/Modules/General/ChannelListViewController/View/ConversationCell.swift
