+++ TinkoffChat/1. PresentationLayer/Modules/General/ChannelViewControler/View/CoversationView.swift
@@ -1,0 +1,19 @@
+//
+//  CoversationView.swift
+//  TinkoffChat
+//
+//  Created by Vera on 12.04.2020.
+//  Copyright © 2020 Vera. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+class CoversationView: UIView{
+    
+    override func awakeFromNib() {
+        super.awakeFromNib()
+    }
+}
+
+
