+++ TinkoffChat/2. ServicesLayer/ConversationCellModel.swift
