+++ TinkoffChat/2. ServicesLayer/MessageCellModel.swift
