+++ TinkoffChat/3. CoreLayer/Modules/Channel/Channel+CoreDataClass.swift
