+++ TinkoffChat/3. CoreLayer/Modules/Channel/Channel+CoreDataProperties.swift
