+++ TinkoffChat/3. CoreLayer/Modules/Message/Message+CoreDataClass.swift
