+++ TinkoffChat/3. CoreLayer/Modules/Message/Message+CoreDataProperties.swift
