+++ TinkoffChat/3. CoreLayer/Modules/StorageManager/StorageProtocol.swift
