+++ TinkoffChat/3. CoreLayer/Modules/User/User+CoreDataClass.swift
