+++ TinkoffChat/3. CoreLayer/Modules/User/User+CoreDataProperties.swift
