--- TinkoffChat/PresentationLayer/Modules/General/ProfileViewController/ProfileModel.swift
@@ -1,13 +1,0 @@
-//
-//  ProfileModel.swift
-//  TinkoffChat
-//
-//  Created by Vera on 12.04.2020.
-//  Copyright © 2020 Vera. All rights reserved.
-//
-
-import Foundation
-
-class ProfileModel {
-    
-}
